Isirmt update

+++ DxlibObjectiveGameLauncher/PlayerScene.cpp
@@ -1,4 +1,5 @@
 #include "PlayerScene.h"
+#include "PopupScene.h"
 
 PlayerScene::PlayerScene()
 	: bg(nullptr), loopButton(nullptr), nowPlaying(nullptr), playBar(nullptr), playListObject(nullptr),
@@ -23,6 +24,16 @@ PlayerScene::PlayerScene(SharingScenes* _sharingScenes)
 		PosVec(50.f, ApplicationPreference::startScenePos + 125.f), PosVec(),
 		"smart20", "123456", ColorPreset::textBlack);
 
+	nextLabel = new TextObject(
+		PosVec(50.f, ApplicationPreference::startScenePos + 200.f), PosVec(),
+		"smart20", "NEXT...", ColorPreset::textBlack);
+	nextTitle = new TextObject(
+		PosVec(70.f, ApplicationPreference::startScenePos + 225.f), PosVec(),
+		"smart20", "NEXT TITLE", ColorPreset::textBlack);
+	nextAuthor = new TextObject(
+		PosVec(70.f, ApplicationPreference::startScenePos + 250.f), PosVec(),
+		"smart20", "NEXT AUTHOR", ColorPreset::textBlack);
+
 	startButton = new ButtonObject(
 		PosVec(
 			(ApplicationPreference::GetBackgroundSize().x - 100.f) / 2.f,
@@ -35,11 +46,14 @@ PlayerScene::PlayerScene(SharingScenes* _sharingScenes)
 		ColorPreset::yellowButtonSelected);
 	startButton->SetInnerAnimation(.1f);
 	startButton->SetOutlineColor(ColorPreset::yellowButtonOuter, 3.f);
+	startButton->SetImageSize(PosVec(75.f, 75.f));
+	startButton->SetImageTurnFlag(false, false);
 
 	playBar = new ProgressObject(
 		PosVec(50.f, ApplicationPreference::GetBackgroundSize().y - 100.f),
 		PosVec(ApplicationPreference::GetBackgroundSize().x - 100.f, 50.f), false);
 	playBar->SetupSlider();
+	playBar->SetEnabledOutline(true);
 	playBar->GetSlider()->SetInnerColor(Color255());
 	playBar->GetSlider()->SetInnerColor(
 		Color255("#73D4C7"));
@@ -61,6 +75,28 @@ PlayerScene::PlayerScene(SharingScenes* _sharingScenes)
 		ColorPreset::tileOuterMouse,
 		ColorPreset::tileOuterMouse,
 		ColorPreset::tileOuterMouse, 2.f);
+	loopButton->SetImageSize(PosVec(75.f, 75.f));
+	loopButton->SetImageTurnFlag(false, false);
+
+	volumeBar = new ProgressObject(
+		PosVec(
+			(ApplicationPreference::GetBackgroundSize().x - 100.f) / 2.f + 100.f,
+			ApplicationPreference::GetBackgroundSize().y - 200.f + 32.5f),
+		PosVec(150.f, 32.5f), false);
+	volumeBar->SetupSlider();
+	volumeBar->SetEnabledOutline(true);
+	volumeBar->GetSlider()->SetInnerColor(Color255());
+	volumeBar->GetSlider()->SetInnerColor(
+		Color255("#73D4C7"));
+	volumeBar->SetInnerColor(Color255(200, 200, 200));
+	volumeBar->SetOutlineColor(Color255(0, 0, 0), 1.f);
+	volumeBar->SetValue(.4f);
+
+	lvolume = new TextObject(
+		PosVec(
+			(ApplicationPreference::GetBackgroundSize().x - 100.f) / 2.f + 100.f,
+			ApplicationPreference::GetBackgroundSize().y - 200.f), PosVec(),
+		"smart20", "VOLUME", ColorPreset::textBlack);
 
 	playListObject = new PlayListObject(
 		PosVec(
@@ -126,14 +162,52 @@ PlayerScene::PlayerScene(SharingScenes* _sharingScenes)
 		2.f);
 	bdeleteList->SetInnerAnimation(.2f);
 
+	/********** JSON 読込 ***********/
+
+	std::stringstream ss;
+	std::ifstream fs;
+	fs.open(ApplicationPreference::setJson, std::ios::binary);
+
+	if (!fs.is_open()) {
+		return;
+	}
+
+	ss << fs.rdbuf();
+	fs.close();
+
+	picojson::value val;
+	ss >> val;
+	std::string err = picojson::get_last_error();
+	if (!err.empty()) {
+		std::cerr << err << std::endl;
+		return;
+	}
+
+	picojson::object& obj = val.get<picojson::object>();
+	picojson::object& jPlayer = obj["Player"].get<picojson::object>();
+
+	ImageChest::CreateImageHandle("Player_play", jPlayer["play"].get<std::string>());
+	ImageChest::CreateImageHandle("Player_stop", jPlayer["stop"].get<std::string>());
+	ImageChest::CreateImageHandle("Player_noloop", jPlayer["noloop"].get<std::string>());
+	ImageChest::CreateImageHandle("Player_oneloop", jPlayer["oneloop"].get<std::string>());
+	ImageChest::CreateImageHandle("Player_loop", jPlayer["loop"].get<std::string>());
+	ImageChest::CreateImageHandle("Player_shuffle", jPlayer["shuffle"].get<std::string>());
+
+	/******** JSON 読込終了 *********/
+
 	layer.AddObject(bg);
 
 	layer.AddObject(songTitle);
 	layer.AddObject(songAuthor);
 	layer.AddObject(songHandleName);
+	layer.AddObject(nextLabel);
+	layer.AddObject(nextTitle);
+	layer.AddObject(nextAuthor);
 	layer.AddObject(startButton);
 	layer.AddObject(loopButton);
 	layer.AddObject(playBar);
+	layer.AddObject(volumeBar);
+	layer.AddObject(lvolume);
 	layer.AddObject(playListObject);
 	layer.AddObject(ideleteList);
 	layer.AddObject(bdeleteList);
@@ -175,7 +249,7 @@ void PlayerScene::Update()
 
 	}
 
-	if (startButton != nullptr)
+	if (startButton != nullptr) {
 		if (startButton->GetMouseSelected()) {
 			startButton->SetMouseOff();
 			if (!MusicPlayer::GetPlaying())
@@ -184,12 +258,44 @@ void PlayerScene::Update()
 				MusicPlayer::StopInList();
 		}
 
-	if (loopButton != nullptr)
+		if (MusicPlayer::GetPlaying()) {
+			startButton->SetImageHandle(ImageChest::GetImageHandle("Player_stop"));
+		}
+		else {
+			startButton->SetImageHandle(ImageChest::GetImageHandle("Player_play"));
+		}
+	}
+
+	if (loopButton != nullptr) {
 		if (loopButton->GetMouseSelected()) {
 			loopButton->SetMouseOff();
 			playStateRotater.Rotate();
 			MusicPlayer::SetPlayingState(playStateRotater.GetNowState());
 		}
+		PlayState playState = MusicPlayer::GetPlayState();
+		switch (playState)
+		{
+		case PlayState::SIMPLE:
+			loopButton->SetImageHandle(ImageChest::GetImageHandle("Player_noloop"));
+			break;
+		case PlayState::ONE_LOOP:
+			loopButton->SetImageHandle(ImageChest::GetImageHandle("Player_oneloop"));
+			break;
+		case PlayState::ALL_LOOP:
+			loopButton->SetImageHandle(ImageChest::GetImageHandle("Player_loop"));
+			break;
+		case PlayState::RANDOM:
+			loopButton->SetImageHandle(ImageChest::GetImageHandle("Player_shuffle"));
+			break;
+		default:
+			break;
+		}
+	}
+
+	if (nextTitle != nullptr)
+		nextTitle->SetText(MusicPlayer::GetPlayNextData().title);
+	if (nextAuthor != nullptr)
+		nextAuthor->SetText(MusicPlayer::GetPlayNextData().author);
 
 	if (playListObject->GetSelectingIndex() >= 0) {
 		playListObject->SetMouseOffIndex();
@@ -211,6 +317,10 @@ void PlayerScene::Update()
 		playBar->SetValue(MusicPlayer::GetPlayingRate());
 	}
 
+	if (volumeBar != nullptr) {
+		MusicPlayer::SetVolumeRate(volumeBar->GetValue());
+	}
+
 	if (playListObject != nullptr) {
 		if (playListObject->SetList(MusicPlayer::GetPlayList(), "mplus40", PosVec(5.f, 5.f))) {
 			canvas->DeleteAllChildren();
@@ -227,14 +337,16 @@ void PlayerScene::Update()
 	}
 
 	if (bdeleteList != nullptr)
-		if (bdeleteList->GetMouseSelected()) {
+		if (bdeleteList->GetMouseSelected() && ideleteList->GetString() != "") {
 			bdeleteList->SetMouseOff();
 			int deleteIndex;
 			try {
 				deleteIndex = std::atoi(ideleteList->GetString().c_str());
 				MusicPlayer::DeleteFromList(deleteIndex);
+				sharingScenes->popupScene->MakeNotice(std::to_string(deleteIndex) + "番をプレイリストから削除しました。");
 			}
 			catch (std::exception e) {
+				sharingScenes->popupScene->MakeNotice("有効な数字を記入してください", "ERROR");
 			}
 			ideleteList->RemakeHandle();
 		}
